new file

+++ bots_commands.py
@@ -15,8 +15,8 @@ def sum_command(update: Update, context: CallbackContext):
     msg = update.message.text
     print (msg)
     items = msg.split()
-    x = int (items[1])
-    y = int (items[2])
+    x = float (items[1])
+    y = float (items[2])
     update.message.reply_text(f'{x} + {y} = {x+y}')
 
 def dif_command(update: Update, context: CallbackContext):
